HistoryTable.tsxに名前変更
