+++ src/scenes/profile/Profile.module.scss
@@ -68,5 +68,21 @@
         text-align: center;
       }
     }
+    .studyHistoryWrapper {
+      width: 100%;
+      .studyHistoryTitle {
+        margin: 80px 0 30px;
+      }
+      .dnnWrapper {
+        padding-bottom: 30px;
+      }
+      .cyberWrapper {
+        padding-bottom: 30px;
+      }
+      .buttonWrapper {
+        color: #e5e5e5;
+        text-align: center;
+      }
+    }
   }
 }
